Added newly discovered blue and also reworked a bunch of the UI
I started reworking the UI to use a button class and a function to check whether a button has been clicked. This makes the code a lot more understandable and easily expanded upon, but it doesn't really have a lot of outwardly visible effects.

Also, I slightly shifted the canvas over for better symmetry around the buttons.

The color blue has been discovered and integrated effectively in this latest version.

This new version contains a mystery function! See if you can find it ;3