Added MyStErY color button
We've added support for a button that, through a proprietary mechanism, generates a random color for you to paint with. Art is about living in the moment and appreciating the unexpected, so we're encouraging this mindset in adding this. We've decided that no further colors are necessary until proven otherwise.

Cursor highlight circle now shows what color you're using. Had to adjust it a bit so as to be visible on any background, but eh. Might try to make it just invert whatever's on the background, but that might involve a modicum of effort.